+++ enumex/__init__.py
@@ -1,20 +1,20 @@
 import sys
 
-__version__ = "3.10.0"
+__version__ = "3.11.0"
 
-if sys.version_info[:2] != (3, 10):
+if sys.version_info[:2] != (3, 11):
     raise RuntimeError(
-        f"This version of enumex is exclusive to Python 3.10.x, "
+        f"This version of enumex is exclusive to Python 3.11, up to patch 8"
         f"and you are using {sys.version_info.major}.{sys.version_info.minor}"
     )
 
 from .enumex import(
-    EnumExMeta,
-    EnumEx, IntEnumEx, FlagEx, IntFlagEx,
+    EnumExType, EnumExMeta,
+    EnumEx, IntEnumEx, StrEnumEx, FlagEx, IntFlagEx, ReprEnumEx,
 )
 
 
 __all__ = [
-        'EnumExMeta',
-        'EnumEx', 'IntEnumEx', 'FlagEx', 'IntFlagEx',
+        'EnumExType', 'EnumExMeta',
+        'EnumEx', 'IntEnumEx', 'StrEnumEx', 'FlagEx', 'IntFlagEx', 'ReprEnumEx',
         ]